chat_list: retire Whitenoise::build_chat_list_item post-Phase 15
Phase 15 (#770) wired session-scoped event dispatch, which removed
the trait-graph depth that originally forced this Whitenoise-side
duplicate. Empirically confirmed: the emit path now delegates to
session.chat_list().build_item(...) without hitting the compiler's
recursion limit.

Drops:
- Whitenoise::build_chat_list_item (~100 lines)
- 3 Whitenoise-side unit tests (DM-peer case ported to
  session/chat_list.rs::tests; the other two were duplicates of
  test_get_chat_list_dm_with_display_name and
  test_get_chat_list_last_message_author_display_name)

Side-benefit: the two implementations diverged on DM-peer resolution
(MDK membership vs SQL projection). Consolidating on the session-side
path retires that divergence — emit and full-list refresh now agree
by construction.
